- new items and power system

+++ app/Dtos/ItemDto.cs
@@ -5,7 +5,7 @@ namespace ClickableGameBackend.Dtos {
     {
         public int Id { get; set; }
         public string? Name { get; set; }
-        public int Power { get; set; }
+        public decimal Power { get; set; }
         public int Count { get; set; }  // Assuming the count of each item a user has
     }
 }
